pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Gustavo Faria <gustavoduartefaria@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
uvicorn = "^0.13.2"
python-dotenv = "^0.15.0"
pydantic = {extras = ["email"], version = "^1.7.3"}
motor = "^2.3.0"
flake8 = "^3.8.4"
pytest-asyncio = "^0.14.0"
pytest-cov = "^2.10.1"
pytest-mock = "^3.4.0"
requests = "^2.25.1"
beautifulsoup4 = "^4.9.3"
google-api-python-client = "^1.12.8"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.8.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "backend.main:start"